API cleanups, added jit_type_name() convenience function

+++ src/var.h
@@ -189,7 +189,7 @@ extern void jitc_var_mask_pop(JitBackend backend);
 extern uint32_t jitc_var_mask_apply(uint32_t index, uint32_t size);
 
 /// Return the default mask
-extern uint32_t jitc_var_mask_default(JitBackend backend, uint32_t size);
+extern uint32_t jitc_var_mask_default(JitBackend backend, size_t size);
 
 /// Start a new scope of the program being recorded
 extern void jitc_new_scope(JitBackend backend);
